JS build convention & fixes
Split up the JS package convention logic into a separate convention plugin, allowing non-kmp packages targeting the JS platform to be configured in a consistent manner
Added a new utility method to access binding values based on binding name for the general binding implementation
Fixed the `sparql:debugging` package not compiling since the binding API change and explicit `utils` dependency
