+++ sparql/common/src/commonMain/kotlin/dev/tesserakt/sparql/Extensions.kt
@@ -1,0 +1,5 @@
+package dev.tesserakt.sparql
+
+import dev.tesserakt.rdf.types.Quad
+
+operator fun Bindings.get(name: String): Quad.Term? = find { it.first == name }?.second
