+++ sparql/debugging/build.gradle.kts
@@ -8,6 +8,8 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
+                implementation(project(":utils"))
+
                 api(project(":sparql:core"))
                 api(project(":sparql:compiler"))
                 api(project(":sparql:runtime"))
